# client/admin-panel/schemas/admin_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Admin",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/admin/auditoriums": {
      "get": {
        "summary": "List Auditoriums",
        "operationId": "list_auditoriums_api_v1_admin_auditoriums_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuditoriumResponse"
                  },
                  "type": "array",
                  "title": "Response List Auditoriums Api V1 Admin Auditoriums Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "summary": "Create Auditorium",
        "operationId": "create_auditorium_api_v1_admin_auditoriums_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuditoriumCreateSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditoriumResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/auditoriums/{auditorium_id}": {
      "get": {
        "summary": "Get Auditorium",
        "operationId": "get_auditorium_api_v1_admin_auditoriums__auditorium_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "auditorium_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Auditorium Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditoriumResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Auditorium",
        "operationId": "update_auditorium_api_v1_admin_auditoriums__auditorium_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "auditorium_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Auditorium Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuditoriumUpdateSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditoriumResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Auditorium",
        "operationId": "delete_auditorium_api_v1_admin_auditoriums__auditorium_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "auditorium_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Auditorium Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/auditoriums/{auditorium_id}/seats": {
      "get": {
        "summary": "Get Seats",
        "operationId": "get_seats_api_v1_admin_auditoriums__auditorium_id__seats_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "auditorium_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Auditorium Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditoriumWithSeats"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Seat",
        "operationId": "create_seat_api_v1_admin_auditoriums__auditorium_id__seats_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "auditorium_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Auditorium Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeatBase"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Replace Seats",
        "operationId": "replace_seats_api_v1_admin_auditoriums__auditorium_id__seats_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "auditorium_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Auditorium Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/SeatBase"
                },
                "title": "Seats"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditoriumWithSeats"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/auditoriums/{auditorium_id}/seats/{seat_id}": {
      "get": {
        "summary": "Get Seat",
        "operationId": "get_seat_api_v1_admin_auditoriums__auditorium_id__seats__seat_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "auditorium_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Auditorium Id"
            }
          },
          {
            "name": "seat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Seat Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Seat",
        "operationId": "update_seat_api_v1_admin_auditoriums__auditorium_id__seats__seat_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "auditorium_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Auditorium Id"
            }
          },
          {
            "name": "seat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Seat Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeatUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Seat",
        "operationId": "delete_seat_api_v1_admin_auditoriums__auditorium_id__seats__seat_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "auditorium_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Auditorium Id"
            }
          },
          {
            "name": "seat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Seat Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/auditoriums/{auditorium_id}/seats/generate": {
      "post": {
        "summary": "Generate Seat Layout",
        "operationId": "generate_seat_layout_api_v1_admin_auditoriums__auditorium_id__seats_generate_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "auditorium_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Auditorium Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateSeatsSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditoriumWithSeats"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuditoriumCreateSchema": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "name",
          "is_active"
        ],
        "title": "AuditoriumCreateSchema"
      },
      "AuditoriumResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "total_capacity": {
            "type": "integer",
            "title": "Total Capacity"
          },
          "is_accessible": {
            "type": "boolean",
            "title": "Is Accessible"
          }
        },
        "type": "object",
        "required": [
          "name",
          "is_active",
          "id",
          "created_at",
          "total_capacity",
          "is_accessible"
        ],
        "title": "AuditoriumResponse"
      },
      "AuditoriumUpdateSchema": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "AuditoriumUpdateSchema"
      },
      "AuditoriumWithSeats": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "total_capacity": {
            "type": "integer",
            "title": "Total Capacity"
          },
          "is_accessible": {
            "type": "boolean",
            "title": "Is Accessible"
          },
          "seats": {
            "items": {
              "$ref": "#/components/schemas/SeatResponse"
            },
            "type": "array",
            "title": "Seats"
          }
        },
        "type": "object",
        "required": [
          "name",
          "is_active",
          "id",
          "created_at",
          "total_capacity",
          "is_accessible",
          "seats"
        ],
        "title": "AuditoriumWithSeats"
      },
      "GenerateSeatsSchema": {
        "properties": {
          "row_count": {
            "type": "integer",
            "title": "Row Count",
            "default": 6
          },
          "seats_per_row": {
            "type": "integer",
            "title": "Seats Per Row",
            "default": 8
          },
          "row_spacing": {
            "type": "integer",
            "title": "Row Spacing",
            "default": 50
          },
          "seat_spacing": {
            "type": "integer",
            "title": "Seat Spacing",
            "default": 45
          },
          "x_offset": {
            "type": "integer",
            "title": "X Offset",
            "default": 60
          },
          "y_offset": {
            "type": "integer",
            "title": "Y Offset",
            "default": 70
          },
          "accessible_rows": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Accessible Rows",
            "default": [
              "A"
            ]
          }
        },
        "type": "object",
        "title": "GenerateSeatsSchema"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SeatBase": {
        "properties": {
          "row": {
            "type": "string",
            "title": "Row"
          },
          "number": {
            "type": "integer",
            "title": "Number"
          },
          "is_available": {
            "type": "boolean",
            "title": "Is Available",
            "default": true
          },
          "is_accessible": {
            "type": "boolean",
            "title": "Is Accessible",
            "default": true
          },
          "x_pos": {
            "type": "integer",
            "title": "X Pos"
          },
          "y_pos": {
            "type": "integer",
            "title": "Y Pos"
          },
          "angle": {
            "type": "integer",
            "title": "Angle",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "row",
          "number",
          "x_pos",
          "y_pos"
        ],
        "title": "SeatBase"
      },
      "SeatResponse": {
        "properties": {
          "row": {
            "type": "string",
            "title": "Row"
          },
          "number": {
            "type": "integer",
            "title": "Number"
          },
          "is_available": {
            "type": "boolean",
            "title": "Is Available",
            "default": true
          },
          "is_accessible": {
            "type": "boolean",
            "title": "Is Accessible",
            "default": true
          },
          "x_pos": {
            "type": "integer",
            "title": "X Pos"
          },
          "y_pos": {
            "type": "integer",
            "title": "Y Pos"
          },
          "angle": {
            "type": "integer",
            "title": "Angle",
            "default": 0
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "auditorium_id": {
            "type": "integer",
            "title": "Auditorium Id"
          }
        },
        "type": "object",
        "required": [
          "row",
          "number",
          "x_pos",
          "y_pos",
          "id",
          "auditorium_id"
        ],
        "title": "SeatResponse"
      },
      "SeatUpdate": {
        "properties": {
          "row": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Row"
          },
          "number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Number"
          },
          "is_available": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Available"
          },
          "is_accessible": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Accessible"
          },
          "x_pos": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "X Pos"
          },
          "y_pos": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Y Pos"
          },
          "angle": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Angle"
          }
        },
        "type": "object",
        "title": "SeatUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}